Added a lvl three + 4
Level three changes it so that it sh... actually I'll keep it a secret. Also debug mode is here, if you can figure out how to turn it on. It's really not difficult. I'll put a menu in so that I can give credit to things I've used, and link to the github.